+++ History.md
@@ -114,6 +114,9 @@
   package on `EmailInternals.NpmModules`. Allow specifying a `MailComposer`
   object to `Email.send` instead of individual options.  #4209
 
+* Expose `Spiderable.requestTimeout` from `spiderable` package to
+  allow apps to set the timeout for running phantomjs.
+
 
 ### Other bug fixes and improvements